[S05P12C104-84] feat /FE_createroom : Add createroom hardcoding

+++ backend/src/main/java/com/ssafy/api/controller/ConferenceController.java
@@ -126,7 +126,7 @@ public class ConferenceController {
         String userId = userDetails.getUsername();
 
         registerPostReq.setUserid(userId);
-        int result = conferenceService.register(registerPostReq);
+        Long result = conferenceService.register(registerPostReq);
 
         return ResponseEntity.ok(ConferenceRes.of(201,"Success",result));
